Use size_t instead of int

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: jmaia <jmaia@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 15:13:14 by jmaia             #+#    #+#             */
-/*   Updated: 2021/11/22 15:14:51 by jmaia            ###   ########.fr       */
+/*   Updated: 2021/11/22 15:17:49 by jmaia            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
+
 size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (*++s)
